+++ Starhaven/Things/Agents/Ghosts/Drone.swift
