+++ Starhaven/Things/Agents/Ghosts/GhostEnums.swift
