+++ Starhaven/Things/Celestial Objects/BlackHole.swift
